+++ packages/components/src/common/keyCodes.js
@@ -12,4 +12,6 @@ export default {
   TAB: 9,
   SPACE: 32,
   BACKSPACE: 8,
+  COMMA: 188,
+  PERIOD: 190,
 };