+++ packages/dynamic-flows/src/flow/layoutService/LayoutService.js
@@ -122,7 +122,7 @@ function convertStepDecisionOption(option) {
   return {
     text: option.description,
     action: {
-      label: option.title,
+      title: option.title,
       method: 'GET',
       url: option.url,
       disabled: option.disabled,
@@ -134,14 +134,13 @@ function convertStepImageToDynamicImage(image) {
   return {
     type: 'image',
     url: image.url,
+    text: image.text,
     margin: 'lg',
   };
 }
 
 function convertStepActionToDynamicAction(action) {
-  const newAction = { ...action, label: action.title };
-  delete newAction.type;
-  delete newAction.title;
+  const newAction = { ...action, title: action.title };
   return {
     type: 'button',
     context: action.type,
@@ -153,7 +152,7 @@ function convertStepReviewToDynamicReview(reviewFields) {
   return {
     type: 'review',
     text: reviewFields.title,
-    definitions: reviewFields.fields.map(convertReviewFieldToDefinition),
+    fields: reviewFields.fields.map(convertReviewFieldToDefinition),
   };
 }
 
@@ -199,7 +198,7 @@ function inlineReferences(layout, schemas, actions, model) {
       return inlineFormSchema(component, schemas, model);
     }
 
-    if (component.type === 'action') {
+    if (component.type === 'button') {
       return inlineAction(component, actions);
     }
 
@@ -229,9 +228,8 @@ function inlineFormSchema(formComponent, schemas, model) {
 }
 
 function inlineAction(actionComponent, actions) {
-  if (actionComponent.$ref) {
-    const newAction = getActionById(actions, actionComponent.$ref);
-    delete newAction.$ref;
+  if (actionComponent.action?.$ref) {
+    const newAction = getActionById(actions, actionComponent.action.$ref);
     return convertStepActionToDynamicAction(newAction);
   }