+++ packages/dynamic-flows/src/ssr.spec.js
@@ -20,6 +20,9 @@ describe('Server side rendering', () => {
   // stick all possible properties we might need to render all components in here
   const allProps = {
     schema: {},
+    components: [],
+    onAction: jest.fn(),
+    onModelChange: jest.fn(),
     onChange: jest.fn(),
     submitted: false,
     name: 'name',