+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,212 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.2.11](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.10...@transferwise/neptune-docs@12.2.11) (2021-06-15)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.10](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.9...@transferwise/neptune-docs@12.2.10) (2021-06-15)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.9](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.8...@transferwise/neptune-docs@12.2.9) (2021-06-14)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.8](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.7...@transferwise/neptune-docs@12.2.8) (2021-06-14)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.7](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.6...@transferwise/neptune-docs@12.2.7) (2021-06-11)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.6](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.5...@transferwise/neptune-docs@12.2.6) (2021-06-11)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.5](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.4...@transferwise/neptune-docs@12.2.5) (2021-06-11)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.4](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.3...@transferwise/neptune-docs@12.2.4) (2021-06-10)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.3](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.2...@transferwise/neptune-docs@12.2.3) (2021-06-10)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.2](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.1...@transferwise/neptune-docs@12.2.2) (2021-06-10)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.2.1](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.2.0...@transferwise/neptune-docs@12.2.1) (2021-06-08)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+# [12.2.0](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.8...@transferwise/neptune-docs@12.2.0) (2021-06-07)
+
+
+### Features
+
+* Add dynamic flow component ([#887](https://github.com/transferwise/neptune-web/issues/887)) ([cbf46d9](https://github.com/transferwise/neptune-web/commit/cbf46d9f670af34888bebd45207921010092cac9))
+
+
+
+
+
+## [12.1.8](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.7...@transferwise/neptune-docs@12.1.8) (2021-06-03)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.7](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.6...@transferwise/neptune-docs@12.1.7) (2021-06-02)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.6](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.5...@transferwise/neptune-docs@12.1.6) (2021-06-01)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.5](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.4...@transferwise/neptune-docs@12.1.5) (2021-05-28)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.4](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.3...@transferwise/neptune-docs@12.1.4) (2021-05-26)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.3](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.2...@transferwise/neptune-docs@12.1.3) (2021-05-26)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.2](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.1...@transferwise/neptune-docs@12.1.2) (2021-05-26)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.1.1](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.1.0...@transferwise/neptune-docs@12.1.1) (2021-05-26)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+# [12.1.0](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.16...@transferwise/neptune-docs@12.1.0) (2021-05-25)
+
+
+### Features
+
+* adding avatar to radio ([#952](https://github.com/transferwise/neptune-web/issues/952)) ([bd0b87a](https://github.com/transferwise/neptune-web/commit/bd0b87a19756e7d356abbc89d8b26df32379ad89))
+
+
+
+
+
+## [12.0.16](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.15...@transferwise/neptune-docs@12.0.16) (2021-05-25)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.0.15](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.14...@transferwise/neptune-docs@12.0.15) (2021-05-25)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.0.14](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.13...@transferwise/neptune-docs@12.0.14) (2021-05-25)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+## [12.0.13](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.12...@transferwise/neptune-docs@12.0.13) (2021-05-24)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
 ## [12.0.12](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@12.0.11...@transferwise/neptune-docs@12.0.12) (2021-05-21)
 
 **Note:** Version bump only for package @transferwise/neptune-docs